Renamed reveiew image gallery to review header gallery
