+++ src/app/(main)/[productType]/components/RatingsModal.tsx
